+++ packages/dnr-rulesets/tasks/build.ts
@@ -12,6 +12,7 @@ import {
     FILTERS_METADATA_FILE_NAME,
     FILTERS_METADATA_I18N_FILE_NAME,
     FILTERS_URL,
+    QUICK_FIXES_FILTER_ID,
     RESOURCES_DIR,
 } from './constants';
 import { getI18nMetadata, getMetadata, type Metadata } from './metadata';
@@ -27,6 +28,7 @@ type FilterDTO = {
 
 /**
  * Gets {@link FilterDTO} array from filter metadata.
+ * AdGuard Quick Fixes filter is excluded from downloading and conversion.
  *
  * @param metadata Filters metadata downloaded from `FILTERS_METADATA_URL`
  * @returns Array of filter data.
@@ -34,11 +36,16 @@ type FilterDTO = {
 const getUrlsOfFiltersResources = async (
     metadata: Metadata,
 ): Promise<FilterDTO[]> => {
-    return metadata.filters.map(({ filterId }) => ({
-        id: filterId,
-        url: `${FILTERS_URL}/${filterId}.txt`,
-        file: `filter_${filterId}.txt`,
-    }));
+    return metadata.filters
+        // We exclude this filter from downloading and conversion,
+        // because it should be loaded from the server on the client and applied
+        // dynamically.
+        .filter(({ filterId }) => filterId !== QUICK_FIXES_FILTER_ID)
+        .map(({ filterId }) => ({
+            id: filterId,
+            url: `${FILTERS_URL}/${filterId}.txt`,
+            file: `filter_${filterId}.txt`,
+        }));
 };
 
 /**