+++ packages/tswebextension/src/lib/common/constants.ts
@@ -1,5 +1,6 @@
 export const USER_FILTER_ID = 0;
 export const ALLOWLIST_FILTER_ID = 100;
+export const QUICK_FIXES_FILTER_ID = 24;
 export const BACKGROUND_TAB_ID = -1;
 export const LF = '\n';